refactor(react): create tree-shaking structure
