+++ packages/react/src/components/feature-flag/index.ts
@@ -1,0 +1,1 @@
+export * from './feature-flag'
