+++ packages/react/src/components/file-upload/file-status.tsx
@@ -9,11 +9,11 @@ import {
 import { fileStatus, type FileStatusVariantProps } from 'styled-system/recipes'
 import { css, cx } from 'styled-system/css'
 import { hstack, vstack } from 'styled-system/patterns'
-import { useCerberusContext } from '../context/cerberus'
-import { ProgressBar, type ProgressBarProps } from './progress/index'
-import { IconButton } from './icon-button/index'
-import { Avatar } from './avatar/avatar'
-import { Field, FieldHelperText } from './field/index'
+import { useCerberusContext } from '../../context/cerberus'
+import { ProgressBar, type ProgressBarProps } from '../progress/index'
+import { IconButton } from '../icon-button/index'
+import { Avatar } from '../avatar/avatar'
+import { Field, FieldHelperText } from '../field/index'
 
 /**
  * This module contains the FileStatus component.
