+++ packages/react/src/components/file-upload/index.ts
@@ -1,0 +1,2 @@
+export * from './file-status'
+export * from './file-uploader'
