+++ packages/react/src/components/for/index.ts
@@ -1,0 +1,1 @@
+export * from './for'
