+++ packages/react/src/components/portal/index.ts
@@ -1,0 +1,1 @@
+export * from './portal'
