+++ packages/react/src/components/show/index.ts
@@ -1,0 +1,1 @@
+export * from './show'
