+++ packages/react/src/components/show/show.tsx
