+++ packages/react/src/components/tag/index.ts
@@ -1,0 +1,1 @@
+export * from './tag'
