+++ packages/react/src/components/text/index.ts
@@ -1,0 +1,1 @@
+export * from './text'
