+++ packages/react/src/index.ts
@@ -14,6 +14,7 @@ export * from './components/deprecated/NavMenuLink'
 export * from './components/deprecated/Legend'
 export * from './hooks/useDate'
 export * from './hooks/useToggle'
+export * from './components/Droppable'
 
 // components
 
@@ -27,28 +28,26 @@ export * from './components/combobox/index'
 export * from './components/cta-dialog/index'
 export * from './components/date-picker/index'
 export * from './components/dialog/index'
-export * from './components/Droppable'
-export * from './components/FeatureFlag'
+export * from './components/feature-flag/index'
 export * from './components/field/index'
 export * from './components/fieldset/index'
-export * from './components/FileStatus'
-export * from './components/FileUploader'
-export * from './components/for'
+export * from './components/file-upload/index'
+export * from './components/for/index'
 export * from './components/icon-button/index'
 export * from './components/menu/index'
 export * from './components/notifications/index'
-export * from './components/Portal'
+export * from './components/portal/index'
 export * from './components/progress/index'
 export * from './components/radio/index'
 export * from './components/rating/index'
 export * from './components/select/index'
-export * from './components/Show'
-export * from './components/Spinner'
+export * from './components/show/index'
+export * from './components/spinner/index'
 export * from './components/switch/index'
 export * from './components/table/index'
 export * from './components/tabs/index'
-export * from './components/Tag'
-export * from './components/Text'
+export * from './components/tag/index'
+export * from './components/text/index'
 export * from './components/toggle/index'
 export * from './components/tooltip/index'
